+++ ic-agent/src/agent/mod.rs
@@ -17,9 +17,9 @@ use agent_error::{HttpErrorPayload, Operation};
 use async_lock::Semaphore;
 use async_trait::async_trait;
 pub use builder::AgentBuilder;
+use bytes::Bytes;
 use cached::{Cached, TimedCache};
-use futures_util::StreamExt;
-use http::{header::CONTENT_TYPE, HeaderMap, Method, StatusCode};
+use http::{header::CONTENT_TYPE, HeaderMap, Method, StatusCode, Uri};
 use ic_ed25519::{PublicKey, SignatureError};
 #[doc(inline)]
 pub use ic_transport_types::{
@@ -28,7 +28,7 @@ pub use ic_transport_types::{
 };
 pub use nonce::{NonceFactory, NonceGenerator};
 use rangemap::{RangeInclusiveMap, RangeInclusiveSet, StepFns};
-use reqwest::{Body, Client, Request, Response};
+use reqwest::{Client, Request, Response};
 use route_provider::{
     dynamic_routing::{
         dynamic_route_provider::DynamicRouteProviderBuilder, node::Node,
@@ -47,6 +47,7 @@ use crate::{
         extract_der, lookup_canister_info, lookup_canister_metadata, lookup_request_status,
         lookup_subnet, lookup_subnet_metrics, lookup_time, lookup_value,
     },
+    agent_error::TransportError,
     export::Principal,
     identity::Identity,
     to_request_id, RequestId,
@@ -67,6 +68,7 @@ use std::{
     fmt::{self, Debug},
     future::{Future, IntoFuture},
     pin::Pin,
+    str::FromStr,
     sync::{Arc, Mutex, RwLock},
     task::{Context, Poll},
     time::Duration,
@@ -1236,14 +1238,25 @@ impl Agent {
         endpoint: &str,
         body: Option<Vec<u8>>,
     ) -> Result<(StatusCode, HeaderMap, Vec<u8>), AgentError> {
-        let create_request_with_generated_url = || -> Result<Request, AgentError> {
+        let body = body.map(Bytes::from);
+
+        let create_request_with_generated_url = || -> Result<http::Request<Bytes>, AgentError> {
             let url = self.route_provider.route()?.join(endpoint)?;
-            let mut http_request = Request::new(method.clone(), url);
-            http_request
-                .headers_mut()
-                .insert(CONTENT_TYPE, "application/cbor".parse().unwrap());
-            *http_request.body_mut() = body.clone().map(Body::from);
-            Ok(http_request)
+            let uri = Uri::from_str(url.as_str())
+                .map_err(|e| AgentError::InvalidReplicaUrl(e.to_string()))?;
+            let body = body.clone().unwrap_or_default();
+            let request = http::Request::builder()
+                .method(method.clone())
+                .uri(uri)
+                .header(CONTENT_TYPE, "application/cbor")
+                .body(body)
+                .map_err(|e| {
+                    AgentError::TransportError(TransportError::Generic(format!(
+                        "unable to create request: {e:#}"
+                    )))
+                })?;
+
+            Ok(request)
         };
 
         let response = self
@@ -1251,40 +1264,13 @@ impl Agent {
             .call(
                 &create_request_with_generated_url,
                 self.max_tcp_error_retries,
+                self.max_response_body_size,
             )
             .await?;
 
-        let http_status = response.status();
-        let response_headers = response.headers().clone();
+        let (parts, body) = response.into_parts();
 
-        // Size Check (Content-Length)
-        if matches!(self
-            .max_response_body_size
-            .zip(response.content_length()), Some((size_limit, content_length)) if content_length > size_limit as u64)
-        {
-            return Err(AgentError::ResponseSizeExceededLimit());
-        }
-
-        let mut body: Vec<u8> = response
-            .content_length()
-            .map_or_else(Vec::new, |n| Vec::with_capacity(n as usize));
-
-        let mut stream = response.bytes_stream();
-
-        while let Some(chunk) = stream.next().await {
-            let chunk = chunk?;
-
-            // Size Check (Body Size)
-            if matches!(self
-                .max_response_body_size, Some(size_limit) if body.len() + chunk.len() > size_limit)
-            {
-                return Err(AgentError::ResponseSizeExceededLimit());
-            }
-
-            body.extend_from_slice(chunk.as_ref());
-        }
-
-        Ok((http_status, response_headers, body))
+        Ok((parts.status, parts.headers, body.to_vec()))
     }
 
     async fn execute(
@@ -1999,13 +1985,89 @@ impl<'agent> IntoFuture for UpdateBuilder<'agent> {
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
 pub trait HttpService: Send + Sync + Debug {
-    /// Perform a HTTP request. Any retry logic should call `req` again, instead of `Request::try_clone`.
+    /// Perform a HTTP request. Any retry logic should call `req` again to get a new request.
     async fn call<'a>(
         &'a self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
         max_retries: usize,
-    ) -> Result<Response, AgentError>;
+        size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError>;
 }
+
+/// Convert from http Request to reqwest's one
+fn from_http_request(req: http::Request<Bytes>) -> Result<Request, AgentError> {
+    let (parts, body) = req.into_parts();
+    let body = reqwest::Body::from(body);
+    // I think it can never fail since it converts from `Url` to `Uri` and `Url` is a subset of `Uri`,
+    // but just to be safe let's handle it.
+    let request = http::Request::from_parts(parts, body)
+        .try_into()
+        .map_err(|e: reqwest::Error| AgentError::InvalidReplicaUrl(e.to_string()))?;
+
+    Ok(request)
+}
+
+/// Convert from reqwests's Response to http one
+#[cfg(not(target_family = "wasm"))]
+async fn to_http_response(
+    resp: Response,
+    size_limit: Option<usize>,
+) -> Result<http::Response<Bytes>, AgentError> {
+    use http_body_util::{BodyExt, Limited};
+
+    let resp: http::Response<reqwest::Body> = resp.into();
+    let (parts, body) = resp.into_parts();
+    let body = Limited::new(body, size_limit.unwrap_or(usize::MAX));
+    let body = body
+        .collect()
+        .await
+        .map_err(|e| {
+            AgentError::TransportError(TransportError::Generic(format!(
+                "unable to read response body: {e:#}"
+            )))
+        })?
+        .to_bytes();
+    let resp = http::Response::from_parts(parts, body);
+
+    Ok(resp)
+}
+
+/// Convert from reqwests's Response to http one.
+/// WASM Response in reqwest doesn't have direct conversion for http::Response,
+/// so we have to hack around using streams.
+#[cfg(target_family = "wasm")]
+async fn to_http_response(
+    resp: Response,
+    size_limit: Option<usize>,
+) -> Result<http::Response<Bytes>, AgentError> {
+    use futures_util::StreamExt;
+    use http_body::Frame;
+    use http_body_util::{Limited, StreamBody};
+
+    // Save headers
+    let status = resp.status();
+    let headers = resp.headers().clone();
+
+    // Convert body
+    let stream = resp.bytes_stream().map(|x| x.map(Frame::data));
+    let body = StreamBody::new(stream);
+    let body = Limited::new(body, size_limit.unwrap_or(usize::MAX));
+    let body = http_body_util::BodyExt::collect(body)
+        .await
+        .map_err(|e| {
+            AgentError::TransportError(TransportError::Generic(format!(
+                "unable to read response body: {e:#}"
+            )))
+        })?
+        .to_bytes();
+
+    let mut resp = http::Response::new(body);
+    *resp.status_mut() = status;
+    *resp.headers_mut() = headers;
+
+    Ok(resp)
+}
+
 #[cfg(not(target_family = "wasm"))]
 #[async_trait]
 impl<T> HttpService for T
@@ -2017,22 +2079,29 @@ where
     #[allow(clippy::needless_arbitrary_self_type)]
     async fn call<'a>(
         mut self: &'a Self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
         max_retries: usize,
-    ) -> Result<Response, AgentError> {
+        size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError> {
         let mut retry_count = 0;
         loop {
-            match Service::call(&mut self, req()?).await {
+            let request = from_http_request(req()?)?;
+
+            match Service::call(&mut self, request).await {
                 Err(err) => {
                     // Network-related errors can be retried.
                     if err.is_connect() {
                         if retry_count >= max_retries {
-                            return Err(AgentError::TransportError(err));
+                            return Err(AgentError::TransportError(TransportError::Reqwest(err)));
                         }
                         retry_count += 1;
                     }
                 }
-                Ok(resp) => return Ok(resp),
+
+                Ok(resp) => {
+                    let resp = to_http_response(resp, size_limit).await?;
+                    return Ok(resp);
+                }
             }
         }
     }
@@ -2048,10 +2117,16 @@ where
     #[allow(clippy::needless_arbitrary_self_type)]
     async fn call<'a>(
         mut self: &'a Self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
-        _: usize,
-    ) -> Result<Response, AgentError> {
-        Ok(Service::call(&mut self, req()?).await?)
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
+        _retries: usize,
+        _size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError> {
+        let request = from_http_request(req()?)?;
+        let response = Service::call(&mut self, request)
+            .await
+            .map_err(|e| AgentError::TransportError(TransportError::Reqwest(e)))?;
+
+        to_http_response(response, _size_limit).await
     }
 }
 
@@ -2065,16 +2140,27 @@ struct Retry429Logic {
 impl HttpService for Retry429Logic {
     async fn call<'a>(
         &'a self,
-        req: &'a (dyn Fn() -> Result<Request, AgentError> + Send + Sync),
+        req: &'a (dyn Fn() -> Result<http::Request<Bytes>, AgentError> + Send + Sync),
         _max_tcp_retries: usize,
-    ) -> Result<Response, AgentError> {
+        _size_limit: Option<usize>,
+    ) -> Result<http::Response<Bytes>, AgentError> {
         let mut retries = 0;
         loop {
             #[cfg(not(target_family = "wasm"))]
-            let resp = self.client.call(req, _max_tcp_retries).await?;
+            let resp = self.client.call(req, _max_tcp_retries, _size_limit).await?;
             // Client inconveniently does not implement Service on wasm
             #[cfg(target_family = "wasm")]
-            let resp = self.client.execute(req()?).await?;
+            let resp = {
+                let request = from_http_request(req()?)?;
+                let resp = self
+                    .client
+                    .execute(request)
+                    .await
+                    .map_err(|e| AgentError::TransportError(TransportError::Reqwest(e)))?;
+
+                to_http_response(resp, _size_limit).await?
+            };
+
             if resp.status() == StatusCode::TOO_MANY_REQUESTS {
                 if retries == 6 {
                     break Ok(resp);